Split admin routes into domain modules
